+++ packages/site/src/utils/snap.ts
@@ -38,6 +38,7 @@ export const createProof = async () => {
   const data = await connector.createProof({
     circuitId: 'credentialAtomicQueryMTPV2OnChain',
     accountAddress: accounts[0],
+    issuerDid: 'did:iden3:tJgV5GSETVoEdg3BeQygWJdNEHHwZTSSiCB1NkM1u',
     query: {
       allowedIssuers: ['*'],
       credentialSubject: {
@@ -49,14 +50,6 @@ export const createProof = async () => {
     },
   });
   console.log(data);
-};
-
-export const createBackup = async () => {
-  await connector.createBackup();
-};
-
-export const recoverBackup = async () => {
-  await connector.recoverBackup();
 };
 
 export const checkStateContractSync = async () => {