+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `@rarimo/rarime`:
+  - Save credentials to ceramic instead of snap store
+
+### Removed
+- `@rarimo/rarime`:
+  - `create` and `recover` backup methods
+- `@rarimo/rarime-connector`:
+  - `create` and `recover` backup methods
 
 ## [0.8.0] - 2023-10-23
 ### Changed